add show selected region and change some id parameter to int
TODO:
change bootstrap default color
shuffle function

+++ app.py
@@ -11,7 +11,20 @@ db.init_app(app)
 
 @app.route('/')
 def index():
-    return render_template('index.html')
+    regions = Region.query.order_by(Region.name).all()
+    region_id = request.args.get('region_id')
+
+    if region_id:
+        pass
+    else:
+        pass
+
+    return render_template('index.html', regions=regions)
+
+
+@app.route('/', methods=['POST'])
+def shuffle_restaurant():
+    pass
 
 
 @app.route('/region', methods=['POST'])
@@ -33,6 +46,7 @@ def create_region():
 @app.route('/region')
 def get_regions():
     regions = Region.query.order_by(Region.name).all()
+
     return render_template('region.html', regions=regions)
 
 
@@ -40,12 +54,13 @@ def get_regions():
 def delete_region():
     name = request.form.get('name')
 
-    if id:
+    if name:
         region = Region.query.filter(Region.name == name).first()
-        for shop in region.shops:
-            db.session.delete(shop)
-        db.session.delete(region)
-        db.session.commit()
+        if region:
+            for shop in region.shops:
+                db.session.delete(shop)
+            db.session.delete(region)
+            db.session.commit()
     else:
         pass
 
@@ -60,9 +75,9 @@ def get_region(id):
 @app.route('/shop', methods=['POST'])
 def create_shop():
     name = request.form.get('name')
-    region_id = request.form.get('region_id')
+    region_id = int(request.form.get('region_id', '0'))
 
-    if region_id == '0' or name == '':
+    if region_id == 0 or name == '':
         pass
     else:
         shop = Shop.query.filter(
@@ -80,14 +95,14 @@ def create_shop():
 @app.route('/shop')
 def get_shops():
     regions = Region.query.order_by(Region.name).all()
-    region_id = request.args.get('region_id')
-    
+    region_id = int(request.args.get('region_id', '0'))
+
     if region_id:
         shops = Shop.query.filter(Shop.region_id == region_id)
     else:
         shops = Shop.query.join(Region).order_by(Region.name).all()
 
-    return render_template('shop.html', regions=regions, shops=shops)
+    return render_template('shop.html', regions=regions, shops=shops, region_id=region_id)
 
 
 @app.route('/shop/delete', methods=['POST'])
@@ -96,8 +111,9 @@ def delete_shop():
 
     if id:
         shop = Shop.query.filter(Shop.id == id).first()
-        db.session.delete(shop)
-        db.session.commit()
+        if shop:
+            db.session.delete(shop)
+            db.session.commit()
     else:
         pass
